+++ tests/FunderMaps.IntegrationTests/Backend/Geocoder/AddressTests.cs
@@ -10,7 +10,7 @@ using Xunit;
 
 namespace FunderMaps.IntegrationTests.Backend.Geocoder
 {
-    // TODO: Use DTO
+    // FUTURE: Use DTO
     public class AddressTests : IClassFixture<AuthBackendWebApplicationFactory>
     {
         private readonly AuthBackendWebApplicationFactory _factory;
