+++ tests/FunderMaps.IntegrationTests/Portal/IncidentPortalTests.cs
@@ -199,7 +199,7 @@ namespace FunderMaps.IntegrationTests.Portal
             Assert.NotNull(returnObject.Name);
         }
 
-        // TODO: FIX: FIXME: XXX
+        // FUTURE: FIX: FIXME: XXX
         [Fact(Skip = "Change testcase to also allow on external address id")]
         public async Task GetRiskAnalysisReturnAnalysis()
         {
